% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{pattern.nlp-package}
\alias{pattern.nlp}
\alias{pattern.nlp-package}
\title{R package to perform sentiment analysis and Parts of Speech tagging for Dutch/French/English/German}
\description{
R package to perform sentiment analysis and Parts of Speech tagging for Dutch/French/English/German.
}

