% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postagging.R
\name{pattern_pos}
\alias{pattern_pos}
\title{POS tagging using the python pattern package including relations.}
\usage{
pattern_pos(x, language, digest = FALSE, as_html = FALSE)
}
\arguments{
\item{x}{a character string in UTF-8}

\item{language}{a character string with possible values 'dutch', 'french', 'english' or 'german'}

\item{digest}{logical indicating to digest::digest the message}

\item{as_html}{logical indicating to return only the xml (for debugging)}
}
\value{
a data.frame with at least the elements sentence.id, sentence.language, chunk.id, chunk.type, chunk.pnp, chunk.relation, 
word.id, word, word.type, word.lemma or an xml object if as_xml is set to TRUE.
}
\description{
POS tagging using the python pattern package including relations. See http://www.clips.ua.ac.be/pattern.
Only dutch/french/english/german
}
\examples{
x <- "Mevrouw wenst zich aan te sluiten bij onze dienst, kan dat wel zomaar?"
pattern_pos(x = x, language = 'dutch')

x <- "Il pleure dans mon coeur comme il pleut sur la ville.
 Quelle est cette langueur qui penetre mon coeur?"
pattern_pos(x = x, language = 'french')

x <- "BNOSAC provides consultancy in open source analytical intelligence. 
 We gather dedicated open source software engineers with a focus on data mining, 
 business intelligence, statistical engineering and advanced artificial intelligence."
pattern_pos(x = x, language = 'english')

x <- "Der Turmer, der schaut zu Mitten der Nacht. 	
 Hinab auf die Graber in Lage
 Der Mond, der hat alles ins Helle gebracht.
 Der Kirchhof, er liegt wie am Tage.
 Da regt sich ein Grab und ein anderes dann."
pattern_pos(x = x, language = 'german')
}

