% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{pattern_sentiment}
\alias{pattern_sentiment}
\title{Sentiment analysis using the python pattern package.}
\usage{
pattern_sentiment(x, language, id = x, digest = FALSE)
}
\arguments{
\item{x}{a character string}

\item{language}{a character string with possible values 'dutch', 'french' or 'english'}

\item{id}{an identifier - to use later on to link}

\item{digest}{logical indicating to digest::digest the message}
}
\value{
a data.frame with elements polarity, subjectivity and id, indicating
the sentiment, subjectivity score and the id
}
\description{
Sentiment analysis using the python pattern package. See http://www.clips.ua.ac.be/pattern.
Only dutch/french/english
}
\examples{
pattern_sentiment("i really really hate iphones", language = "english")
pattern_sentiment("We waren bijna bij de kooien toen er van boven 
 een hoeragejuich losbrak alsof Rudi Vuller door Koeman 
 in z'n kloten was geschopt.", language = "dutch")
pattern_sentiment("j'aime Paris, c'est super", language = "french")
}
