% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{penn_treebank_postags}
\alias{penn_treebank_postags}
\title{POS tags and definitions used in the Penn Treebank.}
\description{
POS tags and definitions used in the Penn Treebank. 
More information: \url{https://www.ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html}
}
\examples{
data(penn_treebank_postags)
str(penn_treebank_postags)
}

