% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{pattern.nlp-package}
\alias{pattern.nlp-package}
\alias{pattern.nlp}
\title{R package to perform sentiment analysis for Dutch/French/English and Parts of Speech tagging for Dutch/French/English/German/Spanish/Italian}
\description{
R package to perform sentiment analysis for Dutch/French/English and Parts of Speech tagging for Dutch/French/English/German/Spanish/Italian.
}
\seealso{
\code{\link{pattern_sentiment}}, \code{\link{pattern_pos}}
}
