% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postagging.R
\name{pattern_pos}
\alias{pattern_pos}
\title{POS tagging using the python pattern package including relations.}
\usage{
pattern_pos(x, language, digest = FALSE, as_html = FALSE, core = FALSE,
  tagset = "penn")
}
\arguments{
\item{x}{a character string in UTF-8}

\item{language}{a character string with possible values 'dutch', 'french', 'english', 'german', 'spanish', 'italian'}

\item{digest}{logical indicating to digest::digest the message}

\item{as_html}{logical indicating to return only the xml (for debugging)}

\item{core}{logical indicating to return only the core fields sentence.id, sentence.language, chunk.id, chunk.type, chunk.pnp, chunk.relation, 
word.id, word, word.type, word.lemma or if deeper level chunks are found, add also these deeper level information are added as columns chunk.level{i}.type/chunk.level{i}.relation/chunkid.level{i} to the data.frame.
Defaults to FALSE, indicating to also add the deeper levels. If set to \code{TRUE}, rbind-ing will be easier as it makes sure the number of columns is always only the core columns. See the examples.}

\item{tagset}{character with the tagset to use. Defaults to 'penn' (the Penn Treebank tagset). Other options are
'universal' for the universal tagset, 'WOTAN' for Dutch, 'parole' for Spanish, 'STTS' for German. For French/Italian, if you did not fill in 'penn' or 'universal' 
it will use the tagset from the Pattern model building phase.}
}
\value{
a data.frame with at least the elements sentence.id, sentence.language, chunk.id, chunk.type, chunk.pnp, chunk.relation, 
word.id, word, word.type, word.lemma or an xml object if \code{as_xml} is set to \code{TRUE}.
Mark that by default all POS tags are mapped on the Penn Treebank tags as available inside this package in \code{\link{penn_treebank_postags}}.
}
\description{
POS tagging using the python pattern package including relations. See http://www.clips.ua.ac.be/pattern.
Only dutch/french/english/german/spanish/italian
}
\examples{
x <- "Dus godvermehoeren met pus in alle puisten, zei die schele van Van Bukburg 
 en hij had nog gelijk ook. Er was toen dat liedje van tietenkonttieten kont tieten kontkontkont, 
 maar dat hoefden we geenseens niet te zingen"
pattern_pos(x = x, language = 'dutch')

x <- "Il pleure dans mon coeur comme il pleut sur la ville.
 Quelle est cette langueur qui penetre mon coeur?"
pattern_pos(x = x, language = 'french')

x <- "BNOSAC provides consultancy in open source analytical intelligence. 
 We gather dedicated open source software engineers with a focus on data mining, 
 business intelligence, statistical engineering and advanced artificial intelligence."
pattern_pos(x = x, language = 'english')

x <- "Der Turmer, der schaut zu Mitten der Nacht. 	
 Hinab auf die Graber in Lage
 Der Mond, der hat alles ins Helle gebracht.
 Der Kirchhof, er liegt wie am Tage.
 Da regt sich ein Grab und ein anderes dann."
pattern_pos(x = x, language = 'german')

x <- "Pasaron cuatro jinetes, sobre jacas andaluzas
 con trajes de azul y verde, con largas capas oscuras."
pattern_pos(x = x, language = 'spanish')

x <- "Avevamo vegliato tutta la notte - i miei amici ed io sotto lampade 
 di moschea dalle cupole di ottone traforato, stellate come le nostre anime, 
 perche come queste irradiate dal chiuso fulgore di un cuore elettrico.
 Avevamo lungamente calpestata su opulenti tappeti orientali la nostra atavica accidia, 
 discutendo davanti ai confini estremi della logica 
 ed annerendo molta carta di frenetiche scritture."
pattern_pos(x = x, language = 'italian')

pattern_pos(x = x, language = 'italian', core = TRUE)
pattern_pos(x = x, language = 'italian', core = FALSE)
pattern_pos(x = x, language = 'italian', as_html = TRUE)
}
